+++ src/packages/core/modal/component/modal-element.element.ts
