+++ src/packages/core/modal/context/modal.context.ts
@@ -1,5 +1,5 @@
 import type { UmbModalConfig, UmbModalType } from './modal-manager.context.js';
-import { UmbModalToken } from './token/modal-token.js';
+import { UmbModalToken } from '../token/modal-token.js';
 import type { IRouterSlot } from '@umbraco-cms/backoffice/external/router-slot';
 import type { UUIModalSidebarSize } from '@umbraco-cms/backoffice/external/uui';
 import { UmbId } from '@umbraco-cms/backoffice/id';
