+++ src/packages/core/modal/index.ts
@@ -1,10 +1,10 @@
-import './modal.element.js';
+import './component/modal.element.js';
 
-export * from './modal-manager.context.js';
-export * from './modal.context.js';
-export * from './modal-route-registration.js';
-export * from './modal-route-registration.controller.js';
+export * from './context/modal-manager.context.js';
+export * from './context/modal.context.js';
+export * from './route-registration/modal-route-registration.js';
+export * from './route-registration/modal-route-registration.controller.js';
 export * from './token/index.js';
-export * from './modal.interfaces.js';
-export * from './modal-element.element.js';
-export * from './modal.element.js';
+export * from './types.js';
+export * from './component/modal-element.element.js';
+export * from './component/modal.element.js';
