+++ src/packages/core/modal/route-registration/modal-route-registration.ts
@@ -1,6 +1,6 @@
-import type { UmbModalContext } from './modal.context.js';
-import type { UmbModalConfig, UmbModalManagerContext } from './modal-manager.context.js';
-import type { UmbModalToken } from './token/modal-token.js';
+import type { UmbModalContext } from '../context/modal.context.js';
+import type { UmbModalConfig, UmbModalManagerContext } from '../context/modal-manager.context.js';
+import type { UmbModalToken } from '../token/modal-token.js';
 import type { IRouterSlot } from '@umbraco-cms/backoffice/external/router-slot';
 import { encodeFolderName } from '@umbraco-cms/backoffice/router';
 import { UmbId } from '@umbraco-cms/backoffice/id';
